Add:UC2_Program to compare float numbers

+++ FindMax.cs
@@ -5,7 +5,7 @@ namespace GenericDemo
     public class FindMax
     {
         
-    public static int MaximumNum(int FirstValue, int SecondValue, int ThirdValue)
+    public static double MaximumNum(double FirstValue, double SecondValue, double ThirdValue)
     {
             if (FirstValue.CompareTo(SecondValue) > 0 && FirstValue.CompareTo(ThirdValue) > 0 ||
                FirstValue.CompareTo(SecondValue) >= 0 && FirstValue.CompareTo(ThirdValue) > 0 ||
